Move out percent change cell component
